fix(intensity): make the recorded map match what the loaders actually store
Three cases where the recorded intensity map and the load-time conversion
disagreed, so the source-domain guarantee failed for supported inputs.

OME-Zarr was excused from record_intensity_domain_map() and recorded as native,
on the grounds that its tile loader copies voxels untouched. It does not. Both
readers instantiate a uint32_t destination and narrow every sample into it --
an explicit static_cast in RawOmezarrLoader::loadTile, an implicit one in the
std::copy of NyxusOmeZarrLoader::loadTile -- so a signed dataset wrapped its
negatives and a real-valued one lost its fraction, and the identity inverse
reported those converted grey levels as source values. The prescan measured the
slide's extrema on the same converted values, so no offset derived from them
could have undone it either. Zarr now behaves like every other backend:
RawOmezarrLoader holds its tile as double and returns the sample as the file
states it; NyxusOmeZarrLoader takes the recorded offset / clamp / dynamic range
and applies the same offset or quantization map grayscale_tiff.h does; the
.zarr early return is gone, so the map is chosen from what the slide holds
rather than from the format it arrived in; and ImageLoader::open derives
quantize / dr / fpmin / fpmax once above the format branch so no backend can
drift from another. A mask is opened with offset 0 and no quantization, which
leaves its labels untouched, and an unsigned store is byte-identical to before.

A constant real-valued slide fell past the quantization branch, which needs
fpmax > fpmin, and kept the default native map -- but the loader has no native
path. ImageLoader::open reads quantize off inten_map, gets false, and hands the
tile loader an offset of 0, which stores trunc(x - 0): a constant 0.5 slide was
stored as grey level 0 and reported as 0, and a constant negative slide lost its
sign the same way. The branch now records the offset map with the unfloored
minimum, so the value itself is the offset and the truncation has nothing left
to eat.

The blank-ROI guard was written as `aux_min == 0 && aux_max == 0`. The offset
map puts a slide's all-pixel minimum on grey level 0, so an ROI lying entirely
at that minimum -- an all-air CT ROI at -1013 HU -- arrives with both at 0 and
cannot be told from an ROI with no pixels. It returned before the histogram ran:
UNIFORMITY kept its default 0 rather than the 1 a single-valued ROI has, ENTROPY
and the percentiles were defaults rather than computations, and the per-bin
HISTOGRAM came back empty, while the affine inverse made the location statistics
look plausible. All four copies of the guard -- in-RAM and out-of-core, 2D and
3D -- now test the pixel cloud for emptiness.

TrivialHistogram::initialize needed the zero-width range to be safe before that
was reachable. The percentile histogram already guarded it with an isnan check
on the bin index; the custom-bin loop did not, and to_grayscale() divides by the
range and casts the resulting NaN to a bin index, which is undefined and only
happens to land on 0 on the toolchains in use. A constant ROI's samples are now
placed in the first bin explicitly, as the percentile histogram has always done.

Tests
-----
* Five analytic cases replace the one that pinned Zarr as native: a real-valued,
  a signed and an unsigned store, and a positive and a negative constant slide.
* signed.ome.zarr (32x32 int16, [-1013, -951]) and float.ome.zarr (32x32
  float32, [-10.5, 5.0]) join the fixtures, uncompressed and a few kB each.
  Five mechanics cases assert the prescan reads the source value and the tile
  loader stores the mapped grey level, with an unsigned negative control.
* A new python mechanics file covers a populated ROI at the slide minimum end to
  end, through both the in-RAM and the out-of-core paths.

+++ src/nyx/raw_omezarr.h
@@ -59,8 +59,11 @@ public:
         else if (dtype_str == "<f8") { data_format_ = 10; fp_pixels_ = true; } //double
         else { data_format_ = 2; } //uint16_t
 
-        // allocate the buffer
-        dest = std::vector<uint32_t> (tile_height_ * tile_width_);
+        // Allocate the buffer. It holds the sample as the file states it: a uint32_t destination
+        // wrapped every negative sample of a signed dataset and truncated every fractional sample
+        // of a real-valued one, so the prescan measured the slide's extrema on converted values and
+        // no offset derived from them could undo the conversion.
+        dest = std::vector<double> (tile_height_ * tile_width_);
 
         // Open the dataset once and cache the handle. The dataset metadata is
         // immutable for the lifetime of this loader, so there is no need to
@@ -126,16 +129,18 @@ public:
     {
     }
 
+    // Only a mask is read through this accessor, and a mask carries non-negative integer labels,
+    // so the narrowing is exact there.
     uint32_t get_uint32_pixel (size_t idx) const
     {
-        uint32_t rv = dest[idx];
+        uint32_t rv = (uint32_t) dest[idx];
         return rv;
     }
 
     double get_dpequiv_pixel (size_t idx) const
     {
-        double rv = (double) dest[idx];
-        return rv;    
+        double rv = dest[idx];
+        return rv;
     }
 
     template<typename FileType>
@@ -163,10 +168,12 @@ public:
         // zero-fill the buffer foreseeing its partial filling at incomplete (tail) tiles
         std::fill(dest.begin(), dest.end(), 0);
         
-        // Copy from buffer to destination tile, handling partial tiles and type conversion
+        // Copy from buffer to destination tile, handling partial tiles. The sample is widened to
+        // double rather than narrowed to an unsigned grey level, so a negative or fractional
+        // dataset reaches the prescan as it is written.
         for (size_t k = 0; k < data_height; ++k) {
             for (size_t j = 0; j < data_width; ++j) {
-                dest[k * tile_width_ + j] = static_cast<uint32_t>(buffer[k * data_width + j]);
+                dest[k * tile_width_ + j] = static_cast<double>(buffer[k * data_width + j]);
             }
         }
     }
@@ -219,6 +226,6 @@ private:
     std::string ds_name_;
     std::unique_ptr<z5::Dataset> ds_;   ///< Cached dataset handle (opened once)
 
-    std::vector<uint32_t> dest;
+    std::vector<double> dest;
 };
 #endif //OMEZARR_SUPPORT
